Move FakeCustomerRepository to test module
